GPUDevice: Add API version field
Also tie shader caches to API version and device LUID. That way we don't
have tons of cache files, and they're regenerated if the GPU/driver
changes.

+++ src/core/host.h
@@ -21,7 +21,7 @@ struct WindowInfo;
 enum class AudioBackend : u8;
 enum class AudioExpansionMode : u8;
 enum class AudioStretchMode : u8;
-enum class RenderAPI : u32;
+enum class RenderAPI : u8;
 class AudioStream;
 class CDImage;
